Revised parsing implementation with a stream-first approach
As requested, the output from parsing now structured based on each stream first
and its associated acitivites, instead of previously where the streams of each
activity were contained in one activity object.

ID's are now zero-indexed and will be guaranteed to be ordered as they are
custom generated from the classes.

Fixed a bug with the time where the minutes may reach 60. A better fix may be
made by someone else in the future.